simplify png src code files structure
