+++ README.md
@@ -2,10 +2,10 @@
 
 workspace repo for the ebony crates.
 
-[<img alt="crates.io" src="https://img.shields.io/crates/v/makura.svg?style=for-the-badge&color=E40046&logo=rust&labelColor=3a3a3a" height="25">](https://crates.io/crates/makura) 
-[<img alt="docs.rs" src="https://img.shields.io/badge/docs.rs-makura-495c9f?style=for-the-badge&logo=docsdotrs&labelColor=3a3a3a" height="25">](https://docs.rs/makura) 
-[<img alt="build:test" src="https://img.shields.io/github/actions/workflow/status/uraneko/makura/rust-ci.yml?branch=main&style=for-the-badge&labelColor=3a3a3a" height="25">](https://github.com/uraneko/makura/actions?query=branch%3Amain)
-[<img alt="license" src="https://img.shields.io/github/license/uraneko/makura?style=for-the-badge&labelColor=3a3a3a&color=ECD53F" height="25">](https://github.com/uraneko/makura/blob/main/LICENSE)
+[<img alt="crates.io" src="https://img.shields.io/crates/v/ebony.svg?style=for-the-badge&color=E40046&logo=rust&labelColor=3a3a3a" height="25">](https://crates.io/crates/ebony) 
+[<img alt="docs.rs" src="https://img.shields.io/badge/docs.rs-ebony-495c9f?style=for-the-badge&logo=docsdotrs&labelColor=3a3a3a" height="25">](https://docs.rs/ebony) 
+[<img alt="build:test" src="https://img.shields.io/github/actions/workflow/status/uraneko/ebony/rust-ci.yml?branch=main&style=for-the-badge&labelColor=3a3a3a" height="25">](https://github.com/uraneko/ebony/actions?query=branch%3Amain)
+[<img alt="license" src="https://img.shields.io/github/license/uraneko/ebony?style=for-the-badge&labelColor=3a3a3a&color=ECD53F" height="25">](https://github.com/uraneko/ebony/blob/main/LICENSE)
 
 ## ToC
 * [Features](#Features)
